+++ tools/hubl/internal/remote.go
@@ -17,7 +17,7 @@ import (
 	"cosmossdk.io/client/v2/autocli/flag"
 
 	"github.com/cosmos/cosmos-sdk/client/flags"
-	clientrpc "github.com/cosmos/cosmos-sdk/client/rpc"
+	"github.com/cosmos/cosmos-sdk/client/grpc/cmtservice"
 	addresscodec "github.com/cosmos/cosmos-sdk/codec/address"
 )
 
@@ -96,7 +96,7 @@ func RemoteCommand(config *Config, configDir string) ([]*cobra.Command, error) {
 		}
 
 		// add comet commands
-		cometCmds := clientrpc.NewCometBFTCommands()
+		cometCmds := cmtservice.NewCometBFTCommands()
 		chainInfo.ModuleOptions[cometCmds.Name()] = cometCmds.AutoCLIOptions()
 
 		appOpts := autocli.AppOptions{